Added a "Populations overview" tab to the ControlView
Building the populations' view works similarly to building the best
individual's phenotype : it is a method to overwrite in the fitness
function (the default implementation returns null, which leads to
nothing being shown). The tab is intended to have an overview of the
distribution of phenotypes in the population / see a graphical
representation of the exploration of the search space.