Improved constant values generation mechanism.
Now distinguishing mutation and generation of a brand new value for
AbstractType (and subsequently nodes holding a value contained in an
AbstractTyp instance, e.g Constant). This permits to have a wide range
for generated values (when initializing population, and later when
regenerating a subtree for instance) and smaller range when mutating a
value.

Added a parameter to Constant nodes to set the probability of value
mutation vs. node regeneration for delegated mutations (see
mutatedCopy()).